+++ client/src/App.css
@@ -1,11 +1,5 @@
 .container {
-    display: flex;
-    flex-direction: column;
-    align-items: center;
+  display: flex;
+  flex-direction: column;
+  align-items: center;
 }
-
-@media screen and (max-width: 800px) {
-    .container{
-      transform: scale(0.9);
-    }
-  }
