+++ client/src/Components/Create/Create.css
@@ -1,66 +1,87 @@
 .container-create {
-    display: flex;
-    flex-direction: column;
-    align-items: center;
+  display: flex;
+  flex-direction: column;
+  align-items: center;
 }
 .container-create .body {
-    border: 2px solid black;
-    border-radius: 4px;
-    background-color: white;
-    box-shadow: rgba(0, 0, 0, 0.35) 0px 5px 15px;
-    width: 400px;
-    height: auto;
-    display: flex;
-    flex-direction: column;
-    gap: 1rem;
-    align-items: center;
-    justify-content: center;
-    padding: 30px 20px;
+  border: 1px solid black;
+  background-color: white;
+  width: 400px;
+  height: auto;
+  display: flex;
+  flex-direction: column;
+  gap: 1rem;
+  align-items: center;
+  justify-content: center;
+  padding: 30px 20px;
 }
-.container-create .body>form{
-    display: flex;
-    flex-direction: column;
-    gap: 1rem;
-    width: 400px;
+.container-create .body > form {
+  display: flex;
+  flex-direction: column;
+  gap: 1rem;
+  width: 400px;
 }
 .container-create .title {
-    flex: 10;
-    display: flex;
-    flex-direction: column;
-    align-items: center;
-    margin: 3rem;
+  flex: 10;
+  display: flex;
+  flex-direction: column;
+  align-items: center;
+  margin: 3rem;
 }
-.container-create .title>span{
-    font-size: 1rem;
-    font-weight: bold;
+.container-create .title > span {
+  font-size: 1rem;
+  font-weight: bold;
 }
-.container-create .body>form label {
-    margin: 3px;
+.container-create .body > form label {
+  margin: 3px;
 }
-.container-create .body>form .user {
-    width: 100%;
-    height: 2rem;
-    outline: none;
-    border: 2px solid black;
-    border-radius: 4px;
-    font-size: 16px;
+.container-create .body > form .user {
+  width: 100%;
+  height: 2rem;
+  outline: none;
+  border: 1px solid black;
+  font-size: 16px;
 }
-.container-create .body>form textarea{
-    max-height: 6rem !important;
+.container-create .body > form textarea {
+  max-height: 6rem !important;
 }
-.container-create .footer .btn{
-    width: 100px;
-    font-size: 18px;
-    padding: 10px 20px;
-    border: none;
-    border-radius: 6px;
-    background-color: cornflowerblue;
-    color: white;
-    cursor: pointer;
+.container-create .footer .btn {
+  width: 100px;
+  font-size: 18px;
+  padding: 10px 20px;
+  border: none;
+  border-radius: 6px;
+  background-color: cornflowerblue;
+  color: white;
+  cursor: pointer;
 }
 
 @media screen and (max-width: 800px) {
-    .container-create {
-        transform: scale(0.9);
-    }
+  .container-create .body {
+    width: 300px;
+  }
+  .container-create .body > form {
+    width: 300px;
+  }
+}
+
+@media screen and (max-width: 400px) {
+  .container-create .body {
+    width: 80%;
+  }
+  .container-create .body > form {
+    font-size: 14px;
+    width: 100%;
+    gap: 10px;
+  }
+  .container-create .body > form label:nth-child(3) {
+    white-space: nowrap;
+  }
+  .container-create .body > form .user  {
+    font-size: 12px;
+  }
+  .container-create .footer .btn {
+    width: 100px;
+    font-size: 15px;
+  }
 }
