+++ client/src/Components/Login/Login.css
@@ -4,14 +4,12 @@
   flex-direction: column;
   align-items: center;
   justify-content: center;
-  border: 2px solid black;
-  border-radius: 4px;
+  border: 1px solid black;
   padding: 10px 10px;
 }
 
 .error-message {
   border: none;
-  border-radius: 4px;
   background-color: rgb(255, 154, 154);
   padding: 5px 10px;
   color: rgb(160, 1, 1);
@@ -29,8 +27,7 @@
 .login-form #email {
   width: 100%;
   height: 2rem;
-  border: 2px solid black;
-  border-radius: 4px;
+  border: 1px solid black;
   font-size: 16px;
   outline: none;
   display: flex;
@@ -39,8 +36,7 @@
 .login-form #password {
   width: 100%;
   height: 2rem;
-  border: 2px solid black;
-  border-radius: 4px;
+  border: 1px solid black;
   font-size: 16px;
   outline: none;
   display: flex;
@@ -53,4 +49,11 @@
   border: none;
   border-radius: 6px;
   cursor: pointer;
+}
+
+@media screen and (max-width: 400px) {
+  .login-container {
+    width: 80%;
+    font-size: 14px;
+  }
 }
