+++ client/src/Components/Main/Main.css
@@ -1,3 +1,3 @@
 .main-page {
-    text-align: center;
+  text-align: center;
 }
