+++ client/src/Components/Modal/ModalDelete/ModalDelete.css
@@ -1,80 +1,94 @@
 .modalDeleteBackground {
-    width: 100%;
-    height: 100%;
-    background-color: rgba(32, 32, 32, 0.5);
-    position: fixed;
-    display: flex;
-    justify-content: center;
-    align-items: center;
-    z-index: 1;
-  }
-  
-  .modalDeleteContainer {
-    width: 400px;
-    height: 350px;
-    border-radius: 4px;
-    border: 2px solid black;
-    background-color: white;
-    box-shadow: rgba(0, 0, 0, 0.35) 0px 5px 15px;
-    display: flex;
-    flex-direction: column;
-    padding: 25px;
-    z-index: 2;
-  }
-  
-  .modalDeleteContainer .title {
-    display: inline-block;
-    text-align: center;
-    margin-top: 10px;
-  }
-  
-  .titleCloseBtn {
-    display: flex;
-    justify-content: flex-end;
-  }
-  
-  .titleCloseBtn button {
-    background-color: transparent;
-    border: none;
-    font-size: 25px;
-    cursor: pointer;
-  }
-  
-  .modalDeleteContainer .body {
-    flex: 50%;
-    display: flex;
-    justify-content: center;
-    align-items: center;
-    font-size: 1.2rem;
-    text-align: center;
-  }
-  
-  .modalDeleteContainer .footer {
-    flex: 20%;
-    display: flex;
-    justify-content: center;
-    align-items: center;
-  }
-  
-  .modalDeleteContainer .footer button {
-    width: 150px;
-    height: 45px;
-    margin: 10px;
-    border: none;
-    background-color: crimson;
-    color: white;
-    border-radius: 8px;
-    font-size: 20px;
-    cursor: pointer;
-  }
-  
-  .modalDeleteContainer .footer #cancelBtn {
-    background-color: cornflowerblue;
-  }
+  width: 100%;
+  height: 100%;
+  background-color: rgba(32, 32, 32, 0.5);
+  position: fixed;
+  display: flex;
+  justify-content: center;
+  align-items: center;
+  z-index: 1;
+}
+
+.modalDeleteContainer {
+  width: 400px;
+  height: 350px;
+  border: 1px solid black;
+  background-color: white;
+  display: flex;
+  flex-direction: column;
+  padding: 25px;
+  z-index: 2;
+}
+
+.modalDeleteContainer .title {
+  display: inline-block;
+  text-align: center;
+  margin-top: 10px;
+}
+
+.titleCloseBtn {
+  display: flex;
+  justify-content: flex-end;
+}
+
+.titleCloseBtn button {
+  background-color: transparent;
+  border: none;
+  font-size: 25px;
+  cursor: pointer;
+}
+
+.modalDeleteContainer .body {
+  flex: 50%;
+  display: flex;
+  justify-content: center;
+  align-items: center;
+  font-size: 1.2rem;
+  text-align: center;
+}
+
+.modalDeleteContainer .footer {
+  flex: 20%;
+  display: flex;
+  justify-content: center;
+  align-items: center;
+}
+
+.modalDeleteContainer .footer button {
+  width: 150px;
+  height: 45px;
+  margin: 10px;
+  border: none;
+  background-color: crimson;
+  color: white;
+  border-radius: 8px;
+  font-size: 20px;
+  cursor: pointer;
+}
+
+.modalDeleteContainer .footer #cancelBtn {
+  background-color: cornflowerblue;
+}
 
 @media screen and (max-width: 800px) {
-  /* I AM LOSING MY FUCKING MIND!!!! */
   .modalDeleteContainer {
-    transform: scale(0.9);
+    width: 300px;
+  }
+}
+
+@media screen and (max-width: 800px) {
+  .modalDeleteContainer {
+    width: 80%;
+  }
+  .modalDeleteContainer .title {
+    font-size: 14px;
+  }
+  .modalDeleteContainer .body {
+    font-size: 14px;
+  }
+  .modalDeleteContainer .footer button {
+    width: 100px;
+    height: 35px;
+    font-size: 15px;
   }
 }
