+++ client/src/Components/Modal/ModalEdit/ModalEdit.css
@@ -9,10 +9,8 @@
   z-index: 1;
 }
 .container-edit {
-  border: 2px solid black;
-  border-radius: 4px;
+  border: 1px solid black;
   background-color: white;
-  box-shadow: rgba(0, 0, 0, 0.35) 0px 5px 15px;
   width: 400px;
   height: auto;
   display: flex;
@@ -42,8 +40,7 @@
   width: 100%;
   height: 2rem;
   outline: none;
-  border: 2px solid black;
-  border-radius: 4px;
+  border: 1px solid black;
 }
 
 .container-edit .body>form textarea {
@@ -74,8 +71,41 @@
 }
 
 @media screen and (max-width: 800px) {
-  /* I AM LOSING MY FUCKING MIND!!!! */
   .container-edit {
-    transform: scale(0.9);
+    width: 300px;
+    gap: 0;
+  }
+  .container-edit .body>form {
+    padding: 0;
+    gap: 0;
+  }
+  .container-edit .body>form label {
+    margin: 10px;
+  }
+}
+
+@media screen and (max-width: 400px) {
+  .container-edit {
+    width: 80%
+  }
+  .container-edit .title {
+    font-size: 14px;
+  }
+  .container-edit .body form {
+    font-size: 14px;
+    width: 100%;
+  }
+  .container-edit .body form label:nth-child(3)>* {
+    display: flex;
+    align-items: center;
+    justify-content: flex-start;
+  }
+  .container-edit .body>form .user {
+    font-size: 12px;
+  }
+  .container-edit .footer button {
+    width: 100px;
+    height: 35px;
+    font-size: 15px;
   }
 }
