+++ client/src/Components/todoList/todoList.css
@@ -1,8 +1,11 @@
+.doc-table {
+  width: auto;
+}
+
 table,
 th,
 td {
   border: 1px solid black;
-  border-radius: 4px;
 }
 table {
   margin: 25px;
@@ -12,22 +15,42 @@ th,
 td {
   padding: 5px 20px;
   border-style: none;
+  max-width: 240px;
+  height: 70px;
 }
 
+td > * {
+  padding: 20px 5px;
+  white-space: nowrap;
+}
+
+td .index, td .status {
+  display: flex;
+  flex-direction: column;
+  align-items: center;
+}
+td .text,
+td .tags {
+  border: 1px solid black;
+  overflow-x: scroll;
+}
 .actionBtns {
   display: flex;
   flex-direction: row;
   align-items: center;
+  justify-content: center;
+  justify-content: center;
 }
 
 .actionBtns button {
   border: none;
-  margin: 1px 5px;
+  margin: 1px 2px;
   border-radius: 8px;
-  padding: 5px 20px;
+  padding: 5px 0;
+  width: 70px;
   color: white;
   background-color: crimson;
-  font-weight: bold;
+  font-size: 15px;
   cursor: pointer;
 }
 
@@ -35,19 +58,93 @@ td {
   background-color: cornflowerblue;
 }
 
-@media screen and (max-width: 800px) {
-  table {
-    transform: scale(0.6);
-    text-align: center;
+@media screen and (max-width: 1400px) {
+  th,
+  td {
+    padding: 5px 10px;
+    max-width: 210px;
+    height: 50px;
+  }
+}
+
+@media screen and (max-width: 1200px) {
+  td > * {
+    white-space: default;
+    overflow-x: scroll;
   }
   th,
   td {
-    padding: 5px 5px;
+    padding: 0px 2px;
+    max-width: 170px;
+    max-height: 50px;
+  }
+  td .date {
+    border: 1px solid black;
+  }
+}
+
+@media screen and (max-width: 1200px) {
+  th,
+  td {
+    max-width: 150px;
+    max-height: 40px;
+  }
+}
+
+@media screen and (max-width: 1000px) {
+  th,
+  td {
+    max-width: 120px;
+    max-height: 40px;
+  }
+
+  td .name {
+    border: 1px solid black;
+    white-space: normal;
+    overflow-x: scroll;
+    overflow-y: scroll;
+  }
+}
+
+@media screen and (max-width: 800px) {
+  th,
+  td {
+    max-width: 100px;
+    font-size: 14px;
+  }
+  tr > :nth-child(6) {
+    display: none;
+  }
+  .actionBtns button {
+    width: 50px;
+    font-size: 12px;
+    padding: 10px 3px;
+  }
+}
+
+@media screen and (max-width: 600px) {
+  th,
+  td {
+    font-size: 12px;
+    padding: 0;
+  }
+  td .text,
+  td .name {
+    max-width: 70px;
+  }
+  table {
+    margin: 10px 0;
+  }
+}
+
+@media screen and (max-width: 500px) {
+  tr > :nth-child(5) {
+    display: none;
   }
 }
 
 @media screen and (max-width: 400px) {
-  table tr .tags {
-    display: none;
+  td .name {
+    max-width: 50px;
   }
 }
